+++ CHANGES.md
@@ -31,6 +31,8 @@
   - @torikizi
 - [ADD] `Sora.Config` に `ClientCert`, `ClientKey`, `CACert` を追加
   - @melpon
+- [ADD] DataChannels に `Header` を追加
+  - @torikizi
 - [ADD] ForwardingFilter に name と priority を追加
   - @torikizi
 - [ADD] ForwardingFilters 機能を使えるようにする